feat(jwt): updated zboxapi client
